requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow
librosa==0.10.1
soundfile==0.12.1
numpy
pydantic==2.5.0
python-multipart==0.0.6
langdetect==1.0.9
pydub==0.25.1
python-dotenv==1.0.0
